requirements.txt
streamlit==1.20.0
msal==1.17.0
requests==2.28.2
azure-storage-blob==12.14.1
